requirements.txt
matplotlib==3.10.3
numpy==2.3.2
Pillow==11.3.0
